Extract capture video code into its own file
