+++ src/capture/capture_video.h
@@ -1,0 +1,32 @@
+/*
+ *  Copyright (C) 2023  The DOSBox Team
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along
+ *  with this program; if not, write to the Free Software Foundation, Inc.,
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+
+#ifndef DOSBOX_CAPTURE_VIDEO_H
+#define DOSBOX_CAPTURE_VIDEO_H
+
+void handle_video_event(bool pressed);
+
+void capture_video(const uint16_t width, const uint16_t height,
+                   const uint8_t bits_per_pixel, const uint16_t pitch,
+                   const uint8_t capture_flags, const float frames_per_second,
+                   const uint8_t* image_data, const uint8_t* palette_data);
+
+void capture_video_add_wave(const uint32_t freq, const uint32_t len,
+                            const int16_t* data);
+
+#endif
