+++ vs/dosbox.vcxproj.filters
@@ -7,6 +7,9 @@
     <ClCompile Include="..\src\capture\capture_image.cpp">
       <Filter>src\capture</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\capture\capture_video.cpp">
+      <Filter>src\capture</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\cpu\callback.cpp">
       <Filter>src\cpu</Filter>
     </ClCompile>
@@ -958,6 +961,9 @@
       <Filter>src\capture</Filter>
     </ClInclude>
     <ClInclude Include="..\src\capture\capture_image.h">
+      <Filter>src\capture</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\capture\capture_video.h">
       <Filter>src\capture</Filter>
     </ClInclude>
     <ClInclude Include="..\src\cpu\core_dynrec\decoder.h">
